+++ src/main/java/io/mattw/jpowder/items/ElementMovement.java
@@ -1,0 +1,5 @@
+package io.mattw.jpowder.items;
+
+public interface ElementMovement {
+    void move(Particle p);
+}
