+++ src/main/java/io/mattw/jpowder/items/Item.java
@@ -1,4 +1,4 @@
-package io.mattw.jpowder;
+package io.mattw.jpowder.items;
 
 import lombok.Getter;
 import lombok.Setter;
